requirements.txt
pandas
numpy
plotly
streamlit
networkx
scipy
python-dotenv
